+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ### v0.2.0 (25 September 2015)
 
-* [ENHANCEMENT] The `ReactCSSTransitionReplace` component require `transitionEnterTimeout` etc. props like `ReactCSSTransitionGroup`
-* [UPGRADE] Upgrade React.js to v0.14.0
+* [DEPENDENCY] Removed the `classnames` dependency. 
+* [ENHANCEMENT] The `${transitionName}-height` class is only added while the height transition is active.
+* [ENHANCEMENT] The `ReactCSSTransitionReplace` component require `transitionEnterTimeout` etc. props like `ReactCSSTransitionGroup`.
+* [UPGRADE] Upgrade React.js to v0.14.0.
 
 ### v0.1.4 (9 September 2015)